+++ EngineTest/Source/Game/Block/Chunk.h
@@ -51,16 +51,16 @@ public:
 	Vector3			GetCenterWs() const { return m_worldBounds.GetCenter(); }
 	AABB3			GetBoundsWs() const { return m_worldBounds; }
 	BlockIterator	GetBlockContainingPosition(const Vector3& worldPosition);
-	Block&			GetBlock(int blockIndex);
+	Block&			GetBlock(uint16 blockIndex);
 	Block&			GetBlock(const IntVector3& blockCoords);
 	IntVector3		GetChunkCoords() const { return m_chunkCoords; }
 	Mesh*			GetMesh() const { return m_mesh; }
 
-	void			SetBlockDefinition(int blockIndex, const BlockDefinition* definition);
+	void			SetBlockDefinition(uint16 blockIndex, const BlockDefinition* definition);
 	void			SetBlockDefinition(const IntVector3& blockCoords, const BlockDefinition* definition);
 
-	static IntVector3	GetBlockCoordsForIndex(int blockIndex);
-	static int			GetBlockIndexForCoords(const IntVector3& coords);
+	static IntVector3	GetBlockCoordsForIndex(uint16 blockIndex);
+	static uint16		GetBlockIndexForCoords(const IntVector3& coords);
 
 
 public:
@@ -73,24 +73,24 @@ public:
 	static constexpr uint8	CHUNK_BITS_Z = 5;											// Number of bits in the Block Index to represent the z coordinate
 	static constexpr uint8	CHUNK_BITS_XZ = CHUNK_BITS_X + CHUNK_BITS_Z;				// Number of bits in the Block Index for both x and z
 
-	static constexpr int	CHUNK_DIMENSIONS_X = (1 << CHUNK_BITS_X);
-	static constexpr int	CHUNK_DIMENSIONS_Y = (1 << CHUNK_BITS_Y);
-	static constexpr int	CHUNK_DIMENSIONS_Z = (1 << CHUNK_BITS_Z);
+	static constexpr uint16	CHUNK_DIMENSIONS_X = (1 << CHUNK_BITS_X);
+	static constexpr uint16	CHUNK_DIMENSIONS_Y = (1 << CHUNK_BITS_Y);
+	static constexpr uint16	CHUNK_DIMENSIONS_Z = (1 << CHUNK_BITS_Z);
 
-	static constexpr int	CHUNK_X_MASK = CHUNK_DIMENSIONS_X - 1;						// Mask that when applied to a BlockIndex will give the block's X coordinate
-	static constexpr int	CHUNK_Y_MASK = (CHUNK_DIMENSIONS_Y - 1) << CHUNK_BITS_XZ;	// Mask that when applied to a BlockIndex will give the block's Y coordinate
-	static constexpr int	CHUNK_Z_MASK = (CHUNK_DIMENSIONS_Z - 1) << CHUNK_BITS_X;	// Mask that when applied to a BlockIndex will give the block's Z coordinate
+	static constexpr uint16	CHUNK_X_MASK = (uint16)CHUNK_DIMENSIONS_X - 1;						// Mask that when applied to a BlockIndex will give the block's X coordinate
+	static constexpr uint16	CHUNK_Y_MASK = (uint16)((CHUNK_DIMENSIONS_Y - 1) << CHUNK_BITS_XZ);	// Mask that when applied to a BlockIndex will give the block's Y coordinate
+	static constexpr uint16	CHUNK_Z_MASK = (uint16)((CHUNK_DIMENSIONS_Z - 1) << CHUNK_BITS_X);	// Mask that when applied to a BlockIndex will give the block's Z coordinate
 
-	static constexpr int	BLOCKS_PER_Z_ROW = CHUNK_DIMENSIONS_X;
-	static constexpr int	BLOCKS_PER_Y_LAYER = CHUNK_DIMENSIONS_X * CHUNK_DIMENSIONS_Z;
-	static constexpr int	BLOCKS_PER_CHUNK = CHUNK_DIMENSIONS_X * CHUNK_DIMENSIONS_Y * CHUNK_DIMENSIONS_Z;
+	static constexpr uint32	BLOCKS_PER_Z_ROW = CHUNK_DIMENSIONS_X;
+	static constexpr uint32	BLOCKS_PER_Y_LAYER = CHUNK_DIMENSIONS_X * CHUNK_DIMENSIONS_Z;
+	static constexpr uint32	BLOCKS_PER_CHUNK = CHUNK_DIMENSIONS_X * CHUNK_DIMENSIONS_Y * CHUNK_DIMENSIONS_Z;
 
 
 private:
 	//-----Private Methods-----
 
 	void PushVerticesForBlock(const IntVector3& blockCoords, const BlockDefinition* def, MeshBuilder& mb);
-	void PushVerticesForBlock(int blockIndex, const BlockDefinition* def, MeshBuilder& mb);
+	void PushVerticesForBlock(uint16 blockIndex, const BlockDefinition* def, MeshBuilder& mb);
 
 
 private:
